+++ glance_widget_android/lib/glance_widget_android.dart
@@ -70,5 +70,6 @@ class GlanceWidgetAndroid extends GlanceWidgetPlatform {
   }
 
   @override
-  Stream<GlanceWidgetAction> get onWidgetAction => _methodChannel.onWidgetAction;
+  Stream<GlanceWidgetAction> get onWidgetAction =>
+      _methodChannel.onWidgetAction;
 }