+++ glance_widget_platform_interface/lib/src/types/widget_action.dart
@@ -49,11 +49,11 @@ class GlanceWidgetAction {
   }
 
   Map<String, dynamic> toMap() => {
-        'widgetId': widgetId,
-        'type': type.name,
-        'payload': payload,
-        'timestamp': timestamp.millisecondsSinceEpoch,
-      };
+    'widgetId': widgetId,
+    'type': type.name,
+    'payload': payload,
+    'timestamp': timestamp.millisecondsSinceEpoch,
+  };
 
   @override
   String toString() =>